pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.5.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tiktok", "config.py"]

[project.scripts]
scrape-post = "tiktok.scrape_post:main"
scrape-comments = "tiktok.scrape_comments:main"
download-media = "tiktok.download_media:main"
analyze-video = "tiktok.analyze_video:main"
tiktok-pipeline = "tiktok.pipeline:main"
